+++ tests/PSRule.Tests/PSRule.Common.Tests.ps1
@@ -284,19 +284,29 @@ Describe 'Invoke-PSRule' -Tag 'Invoke-PSRule','Common' {
             Name = 'TestObject1'
             Value = 1
         }
+        $testObject.PSObject.TypeNames.Insert(0, 'TestType');
 
         It 'Yaml' {
             $result = @($testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Name 'FromFile1' -OutputFormat Yaml);
             $result | Should -Not -BeNullOrEmpty;
             $result | Should -BeOfType System.String;
-            $result -match 'ruleName: FromFile1' | Should -Be $True;
+            $result -cmatch 'ruleName: FromFile1' | Should -Be $True;
+            $result -cmatch 'outcome: Pass' | Should -Be $True;
+            $result -cmatch 'targetName: TestObject1' | Should -Be $True;
+            $result -cmatch 'targetType: TestType' | Should -Be $True;
+            $result | Should -Match 'tag:(\r|\n){1,2}\s{2,}(test: Test1|category: group1)';
+            $result | Should -Not -Match 'targetObject:';
         }
 
         It 'Json' {
             $result = @($testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Name 'FromFile1' -OutputFormat Json);
             $result | Should -Not -BeNullOrEmpty;
             $result | Should -BeOfType System.String;
-            $result -match '"ruleName":"FromFile1"' | Should -Be $True;
+            $result -cmatch '"ruleName":"FromFile1"' | Should -Be $True;
+            $result -cmatch '"outcome":"Pass"' | Should -Be $True;
+            $result -cmatch '"targetName":"TestObject1"' | Should -Be $True;
+            $result -cmatch '"targetType":"TestType"' | Should -Be $True;
+            $result | Should -Not -Match '"targetObject":';
         }
     }
 
@@ -1003,6 +1013,23 @@ Describe 'New-PSRuleOption' -Tag 'Option','Common','New-PSRuleOption' {
         It 'from YAML' {
             $option = New-PSRuleOption -Option (Join-Path -Path $here -ChildPath 'PSRule.Tests.yml');
             $option.Logging.RulePass | Should -Be Warning;
+        }
+    }
+
+    Context 'Read Output.As' {
+        It 'from default' {
+            $option = New-PSRuleOption;
+            $option.Output.As | Should -Be 'Detail';
+        }
+
+        It 'from Hashtable' {
+            $option = New-PSRuleOption -Option @{ 'Output.As' = 'Summary' };
+            $option.Output.As | Should -Be Summary;
+        }
+
+        It 'from YAML' {
+            $option = New-PSRuleOption -Option (Join-Path -Path $here -ChildPath 'PSRule.Tests.yml');
+            $option.Output.As | Should -Be Summary;
         }
     }